requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
faiss-cpu==1.7.4
numpy==1.24.3
python-multipart==0.0.6
tqdm==4.66.1
